+++ r8s/r8scli/service/adapter_client.py
@@ -844,6 +844,34 @@ class AdapterClient:
                                    method=HTTP_GET,
                                    payload=request)
 
+    def resource_group_get(self, parent_id: str):
+        request = {
+            PARAM_PARENT_ID: parent_id
+        }
+
+        return self.__make_request(resource=API_RESOURCE_GROUPS,
+                                   method=HTTP_GET,
+                                   payload=request)
+
+    def resource_group_patch(self, parent_id: str, add_tags: list = None,
+                             remove_tags: list = None,
+                             add_resource_groups: list = None,
+                             remove_resource_groups: list = None):
+        request = {
+            PARAM_PARENT_ID: parent_id,
+            PARAM_ADD_TAGS: add_tags,
+            PARAM_REMOVE_TAGS: remove_tags,
+            PARAM_ADD_RESOURCE_GROUPS: add_resource_groups,
+            PARAM_REMOVE_RESOURCE_GROUPS: remove_resource_groups
+        }
+
+        request = {k: v for k, v in request.items() if
+                   v and isinstance(v, list)}
+
+        return self.__make_request(resource=API_RESOURCE_GROUPS,
+                                   method=HTTP_PATCH,
+                                   payload=request)
+
     def shape_get(self, name=None, cloud=None):
         request = {}
         if name: