+++ src/services/rbac/endpoint_to_permission_mapping.py
@@ -99,6 +99,10 @@ ENDPOINT_PERMISSION_MAPPING = {
         PATCH_METHOD: 'r8s:parent:update_shape_rule',
         DELETE_METHOD: 'r8s:parent:remove_shape_rule',
     },
+    '/parents/resource-groups/': {
+        GET_METHOD: 'r8s:parent:describe_group_config',
+        PATCH_METHOD: 'r8s:parent:update_group_config',
+    },
     '/parents/shape-rules/dry-run/': {
         GET_METHOD: 'r8s:parent:dry_run_shape_rule'
     },