+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.12.0] - 2025-02-06
+* Extend resource group support:
+  * Add `r8s parent group` cli group:
+    * `add_allowed_tags` - Adds specified tags to resource group configuration
+    * `remove_tags` - Removes specified resource groups from group configuration
+    * `add_resource_group` - Adds specified resource groups to configuration
+    * `remove_resource_group` - Removes specified resource groups from configuration
+
+
 ## [3.11.0] - 2024-09-17
 * Update onprem version of the service
 * Add readable recommendation text to DefectDojo findings